+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* Support multiple ongoing activities.
+
 ## [2.0.0-alpha.1](https://crates.io/crates/rtw/2.0.0-alpha.1) Jul 12, 2020
 
 * bump `htp` to 0.2.1 (fix `next <weekday>`)